Refactored camel cdi classes into separate flowable-camel-cdi module
