+++ modules/flowable-camel-cdi/src/main/java/org/flowable/camel/cdi/impl/CdiCamelBehaviorDefaultImpl.java
@@ -11,22 +11,23 @@
  * limitations under the License.
  */
 
-package org.flowable.camel.impl;
+package org.flowable.camel.cdi.impl;
 
-import org.flowable.camel.CdiCamelBehavior;
 import org.flowable.camel.FlowableEndpoint;
+import org.flowable.camel.cdi.CdiCamelBehavior;
+import org.flowable.camel.impl.CamelBehaviorDefaultImpl;
 
 /**
- * CDI equivalent of {@link CdiCamelBehaviorCamelBodyImpl}
+ * Cdi equivalent of {@link CamelBehaviorDefaultImpl}
  * 
  * @author Zach Visagie
  */
-public class CdiCamelBehaviorCamelBodyImpl extends CdiCamelBehavior {
+public class CdiCamelBehaviorDefaultImpl extends CdiCamelBehavior {
 
     private static final long serialVersionUID = 1L;
 
     @Override
     protected void setPropertTargetVariable(FlowableEndpoint endpoint) {
-        toTargetType = TargetType.BODY;
+        toTargetType = TargetType.PROPERTIES;
     }
 }
