+++ modules/flowable-camel-cdi/src/test/java/org/flowable/camel/cdi/BaseCamelCdiFlowableTestCase.java
