+++ modules/flowable-camel-cdi/src/test/java/org/flowable/camel/cdi/classpath.txt
@@ -1,0 +1,92 @@
+jackson-core-2.9.3.jar
+jackson-databind-2.9.3.jar
+jackson-annotations-2.9.3.jar
+commons-logging-1.2.jar
+java-uuid-generator-3.1.3.jar
+log4j-1.2.13.jar
+commons-email-1.5.jar
+javax.mail-1.5.6.jar
+commons-codec-1.10.jar
+commons-lang3-3.7.jar
+mybatis-3.4.1.jar
+spring-beans-4.3.9.RELEASE.jar
+spring-core-4.3.9.RELEASE.jar
+joda-time-2.9.9.jar
+slf4j-api-1.7.25.jar
+jcl-over-slf4j-1.7.25.jar
+spring-context-4.3.9.RELEASE.jar
+spring-jdbc-4.3.9.RELEASE.jar
+spring-tx-4.3.9.RELEASE.jar
+spring-orm-4.3.9.RELEASE.jar
+commons-dbcp-1.4.jar
+commons-pool-1.5.4.jar
+camel-spring-2.17.1.jar
+spring-aop-4.3.9.RELEASE.jar
+spring-expression-4.3.9.RELEASE.jar
+jaxb-impl-2.2.11.jar
+camel-cdi-2.17.1.jar
+camel-core-2.17.1.jar
+jaxb-core-2.2.11.jar
+arquillian-junit-container-1.0.2.Final.jar
+arquillian-junit-core-1.0.2.Final.jar
+arquillian-test-api-1.0.2.Final.jar
+arquillian-core-api-1.0.2.Final.jar
+arquillian-test-spi-1.0.2.Final.jar
+arquillian-core-spi-1.0.2.Final.jar
+arquillian-container-test-api-1.0.2.Final.jar
+shrinkwrap-api-1.0.1.jar
+arquillian-container-test-spi-1.0.2.Final.jar
+arquillian-core-impl-base-1.0.2.Final.jar
+arquillian-test-impl-base-1.0.2.Final.jar
+arquillian-container-impl-base-1.0.2.Final.jar
+arquillian-config-api-1.0.2.Final.jar
+arquillian-config-impl-base-1.0.2.Final.jar
+shrinkwrap-descriptors-spi-2.0.0-alpha-3.jar
+arquillian-container-test-impl-base-1.0.2.Final.jar
+shrinkwrap-impl-base-1.0.0-cr-1.jar
+shrinkwrap-spi-1.0.0-cr-1.jar
+junit-4.12.jar
+hamcrest-core-1.3.jar
+h2-1.4.196.jar
+jboss-annotations-api_1.1_spec-1.0.1.Final.jar
+jboss-connector-api_1.6_spec-1.0.0.Final.jar
+jboss-ejb-api_3.1_spec-1.0.0.CR2.jar
+jaxrpc-api-1.1.jar
+jboss-el-api_2.2_spec-1.0.0.Final.jar
+jboss-jsf-api_2.0_spec-1.0.0.Final.jar
+jboss-interceptors-api_1.1_spec-1.0.0.Beta1.jar
+jboss-jad-api_1.2_spec-1.0.0.Final.jar
+jboss-j2eemgmt-api_1.1_spec-1.0.0.Final.jar
+jboss-jaxr-api_1.0_spec-1.0.0.Final.jar
+jboss-jms-api_1.1_spec-1.0.0.Final.jar
+jboss-jaspi-api_1.0_spec-1.0.0.Final.jar
+jboss-jacc-api_1.4_spec-1.0.0.Final.jar
+jboss-jsp-api_2.2_spec-1.0.0.Final.jar
+jboss-jstl-api_1.2_spec-1.0.1.Final.jar
+jboss-servlet-api_3.0_spec-1.0.0.Beta2.jar
+jboss-transaction-api_1.1_spec-1.0.0.Final.jar
+jboss-jaxrs-api_1.1_spec-1.0.0.Final.jar
+jboss-jaxb-api_2.2_spec-1.0.2.Final.jar
+jboss-jaxrpc-api_1.1_spec-1.0.0.Final.jar
+jboss-saaj-api_1.3_spec-1.0.1.Final.jar
+jboss-jaxws-api_2.2_spec-2.0.0.Final.jar
+activation-1.1.jar
+cdi-api-1.0-SP4.jar
+javax.inject-1.jar
+jsr181-api-1.0-MR1.jar
+mail-1.4.4.jar
+hibernate-jpa-2.0-api-1.0.1.Final.jar
+validation-api-1.0.0.GA.jar
+arquillian-weld-ee-embedded-1.1-1.0.0.CR9.jar
+arquillian-container-spi-1.0.2.Final.jar
+shrinkwrap-descriptors-api-base-2.0.0-alpha-3.jar
+arquillian-testenricher-cdi-1.1.4.Final.jar
+weld-core-1.1.33.Final.jar
+weld-spi-1.1.Final.jar
+jsr250-api-1.0.jar
+javassist-3.18.1-GA.jar
+guava-17.0.jar
+slf4j-ext-1.7.2.jar
+cal10n-api-0.7.7.jar
+weld-api-1.1.Final.jar
+slf4j-simple-1.7.2.jar
