+++ modules/flowable-camel-cdi/src/test/java/org/flowable/camel/cdi/named/CdiCustomContextTest.java
@@ -10,7 +10,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.flowable.camel.cdi.std;
+package org.flowable.camel.cdi.named;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNull;
@@ -20,6 +20,7 @@ import java.util.List;
 import java.util.Map;
 
 import javax.inject.Inject;
+import javax.inject.Named;
 
 import org.apache.camel.CamelContext;
 import org.apache.camel.Exchange;
@@ -28,7 +29,6 @@ import org.apache.camel.Route;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.flowable.camel.FlowableProducer;
-import org.flowable.camel.SimpleProcessTest;
 import org.flowable.cdi.impl.util.ProgrammaticBeanLookup;
 import org.flowable.engine.runtime.ProcessInstance;
 import org.flowable.engine.test.Deployment;
@@ -36,13 +36,14 @@ import org.junit.After;
 import org.junit.Test;
 
 /**
- * Adapted from {@link SimpleProcessTest}.
+ * Adapted from {@link CustomContextTest}.
  * 
  * @author Zach Visagie
  */
-public class CdiSimpleProcessTest extends StdCamelCdiFlowableTestCase {
+public class CdiCustomContextTest extends NamedCamelCdiFlowableTestCase {
 
     @Inject
+    @Named("camelContext")
     protected CamelContext camelContext;
 
     protected MockEndpoint service1;
@@ -51,21 +52,23 @@ public class CdiSimpleProcessTest extends StdCamelCdiFlowableTestCase {
 
     @Override
     public void setUp() throws Exception {
-    	super.setUp();
-        service1 = (MockEndpoint) camelContext.getEndpoint("mock:service1");
-        service1.reset();
-        service2 = (MockEndpoint) camelContext.getEndpoint("mock:service2");
-        service2.reset();
+        super.setUp();
         camelContext.addRoutes(new RouteBuilder() {
 
             @Override
             public void configure() throws Exception {
                 from("direct:start").to("flowable:camelProcess");
-                from("flowable:camelProcess:serviceTask1").setBody().exchangeProperty("var1").to("mock:service1").setProperty("var2").constant("var2").setBody().properties();
-                from("direct:receive").to("flowable:camelProcess:receive");
+                from("flowable:camelProcess:serviceTask1").setBody().exchangeProperty("var1").to("mock:service1").setProperty("var2").constant("var2").setBody()
+                                .properties();
                 from("flowable:camelProcess:serviceTask2?copyVariablesToBodyAsMap=true").to("mock:service2");
+                from("direct:receive").to("flowable:camelProcess:receive");
             }
         });
+
+        service1 = (MockEndpoint) camelContext.getEndpoint("mock:service1");
+        service1.reset();
+        service2 = (MockEndpoint) camelContext.getEndpoint("mock:service2");
+        service2.reset();
     }
 
     @After
@@ -78,19 +81,20 @@ public class CdiSimpleProcessTest extends StdCamelCdiFlowableTestCase {
     }
 
     @Test
-    @Deployment(resources = { "process/example.bpmn20.xml" })
+    @Deployment(resources = { "process/custom.bpmn20.xml" })
     public void testRunProcess() throws Exception {
-        CamelContext ctx = ProgrammaticBeanLookup.lookup(CamelContext.class);
+        CamelContext ctx = (CamelContext) ProgrammaticBeanLookup.lookup("camelContext");
         ProducerTemplate tpl = ctx.createProducerTemplate();
         service1.expectedBodiesReceived("ala");
 
         Exchange exchange = ctx.getEndpoint("direct:start").createExchange();
         exchange.getIn().setBody(Collections.singletonMap("var1", "ala"));
         tpl.send("direct:start", exchange);
-        String instanceId = (String) exchange.getProperty("PROCESS_ID_PROPERTY");
 
+        String instanceId = (String) exchange.getProperty("PROCESS_ID_PROPERTY");
+        List<ProcessInstance> processInstances = processEngine.getRuntimeService().createProcessInstanceQuery().list();
         ProcessInstance processInstance = processEngine.getRuntimeService().createProcessInstanceQuery().processInstanceId(instanceId).singleResult();
-        assertEquals(false,processInstance.isEnded());
+        assertEquals(false, processInstance.isEnded());
 
         tpl.sendBodyAndProperty("direct:receive", null, FlowableProducer.PROCESS_ID_PROPERTY, instanceId);
 
@@ -99,9 +103,10 @@ public class CdiSimpleProcessTest extends StdCamelCdiFlowableTestCase {
         assertNull(processInstance);
 
         service1.assertIsSatisfied();
-        Map<?, ?> m = service2.getExchanges().get(0).getIn().getBody(Map.class);
+
+        @SuppressWarnings("rawtypes")
+        Map m = service2.getExchanges().get(0).getIn().getBody(Map.class);
         assertEquals("ala", m.get("var1"));
         assertEquals("var2", m.get("var2"));
     }
-
 }
