+++ modules/flowable-camel-cdi/src/test/java/org/flowable/camel/cdi/named/NamedCamelCdiFlowableTestCase.java
