+++ modules/flowable-camel-cdi/src/test/java/org/flowable/camel/cdi/std/CdiCamelVariableBodyTest.java
@@ -25,8 +25,7 @@ import org.apache.camel.LoggingLevel;
 import org.apache.camel.Route;
 import org.apache.camel.builder.RouteBuilder;
 import org.apache.camel.component.mock.MockEndpoint;
-import org.flowable.camel.CamelVariableBodyTest;
-import org.flowable.camel.impl.CdiCamelBehaviorCamelBodyImpl;
+import org.flowable.camel.cdi.impl.CdiCamelBehaviorCamelBodyImpl;
 import org.flowable.engine.runtime.ProcessInstance;
 import org.flowable.engine.test.Deployment;
 import org.flowable.task.api.Task;
