+++ modules/flowable-camel-cdi/src/test/resources/META-INF/services/org.flowable.cdi.spi.ProcessEngineLookup
