+++ modules/flowable-camel-cdi/src/test/resources/flowable.cfg.xml
