+++ modules/flowable-camel-cdi/src/test/resources/log4j.properties
@@ -1,0 +1,6 @@
+log4j.rootLogger=INFO, CA
+
+# ConsoleAppender
+log4j.appender.CA=org.apache.log4j.ConsoleAppender
+log4j.appender.CA.layout=org.apache.log4j.PatternLayout
+log4j.appender.CA.layout.ConversionPattern= %d{hh:mm:ss,SSS} [%t] %-5p %c %x - %m%n
