+++ modules/flowable-camel-cdi/src/test/resources/process/asyncPing.bpmn20.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<definitions id="definitions"
+             xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL"
+             xmlns:activiti="http://activiti.org/bpmn"
+             targetNamespace="Examples" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+             xsi:schemaLocation="
+             http://www.omg.org/spec/BPMN/20100524/MODEL http://www.omg.org/spec/BPMN/2.0/20100501/BPMN20.xsd">
+
+
+    <process id="asyncPingProcess">
+
+        <startEvent id="start"/>
+        <sequenceFlow sourceRef="start" targetRef="serviceAsyncPing"/>
+        <serviceTask id="serviceAsyncPing" activiti:type="camel" activiti:async="true"/>
+        <sequenceFlow sourceRef="serviceAsyncPing" targetRef="receiveAsyncPing"/>
+        <receiveTask id="receiveAsyncPing" name="Wait State" />
+        <sequenceFlow sourceRef="receiveAsyncPing" targetRef="end"/>
+        <endEvent id="end"/>
+
+    </process>
+
+</definitions>
