+++ modules/flowable-camel-cdi/src/test/resources/process/customNonDefaultName.bpmn20.xml
